feat: re-organise clients to align with Data.gov.sg's categories
