+++ datagovsg/economy/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2019 Yuhui
+# Copyright 2019-2025 Yuhui
 #
 # Licensed under the GNU General Public License, Version 3.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,25 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from datetime import datetime
-
-from .ckan import Client as Ckan
-from .economy import Client as Economy
-from .environment import Client as Environment
-from .housing import Client as Housing
-from .transport import Client as Transport
-
-NAME = 'datagovsg'
-VERSION = '2.0.0' # Production
-VERSION = f'{VERSION}.{datetime.now().strftime("%Y%m%d%H%M")}' # Development
-AUTHOR = 'Yuhui'
-AUTHOR_EMAIL = 'yuhuibc@gmail.com'
+from .client import Client
 
 __all__ = [
-    'Ckan',
-    'Economy',
-    'Environment',
-    'Housing',
-    'Transport',
+    'Client',
 ]
-__version__ = VERSION
