+++ datagovsg/economy/constants.py
@@ -1,4 +1,4 @@
-# Copyright 2019 Yuhui
+# Copyright 2019-2025 Yuhui
 #
 # Licensed under the GNU General Public License, Version 3.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,13 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Constants for all Technology-related APIs."""
+"""Constants for all Economy-related APIs."""
 
 from ..constants import BASE_V1_API_ENDPOINT
 
-TECHNOLOGY_API_ENDPOINT = f'{BASE_V1_API_ENDPOINT}/technology'
+ECONOMY_API_ENDPOINT = f'{BASE_V1_API_ENDPOINT}/technology'
 
-IPOS_API_ENDPOINT = f'{TECHNOLOGY_API_ENDPOINT}/ipos'
+IPOS_API_ENDPOINT = f'{ECONOMY_API_ENDPOINT}/ipos'
 
 IPOS_DESIGNS_API_ENDPOINT = f'{IPOS_API_ENDPOINT}/designs'
 IPOS_PATENTS_API_ENDPOINT = f'{IPOS_API_ENDPOINT}/patents'
