+++ datagovsg/economy/types_args.py
@@ -1,4 +1,4 @@
-# Copyright 2019 Yuhui
+# Copyright 2025 Yuhui. All rights reserved.
 #
 # Licensed under the GNU General Public License, Version 3.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,8 +12,26 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .client import Client
+"""Data.gov.sg custom types for Economy client methods' arguments."""
+
+from datetime import date
+from typing import NotRequired
+try:
+    from typing import TypedDict
+except ImportError:
+    TypedDict = dict
+
+class EconomyArgsDict(TypedDict):
+    """Type definition for designs(), patents() and trademarks() input \
+        arguments
+    """
+
+    lodgement_date: NotRequired[date]
+    """Retrieve applications lodged at that moment in time.
+
+    :example: date(2025, 1, 12)
+    """
 
 __all__ = [
-    'Client',
+    'EconomyArgsDict',
 ]
