+++ datagovsg/housing/__init__.py
@@ -1,4 +1,4 @@
-# Copyright 2025 Yuhui. All rights reserved.
+# Copyright 2025 Yuhui
 #
 # Licensed under the GNU General Public License, Version 3.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,26 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Data.gov.sg custom types for Technology client methods' arguments."""
-
-from datetime import date
-from typing import NotRequired
-try:
-    from typing import TypedDict
-except ImportError:
-    TypedDict = dict
-
-class TechnologyArgsDict(TypedDict):
-    """Type definition for designs(), patents() and trademarks() input \
-        arguments
-    """
-
-    lodgement_date: NotRequired[date]
-    """Retrieve applications lodged at that moment in time.
-
-    :example: date(2025, 1, 12)
-    """
+from .client import Client
 
 __all__ = [
-    'TechnologyArgsDict',
+    'Client',
 ]
