+++ datagovsg/housing/constants.py
@@ -1,4 +1,4 @@
-# Copyright 2019 Yuhui
+# Copyright 2025 Yuhui
 #
 # Licensed under the GNU General Public License, Version 3.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,25 +12,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from datetime import datetime
+"""Constants for all Housing-related APIs."""
 
-from .ckan import Client as Ckan
-from .economy import Client as Economy
-from .environment import Client as Environment
-from .housing import Client as Housing
-from .transport import Client as Transport
+from ..constants import BASE_V1_API_ENDPOINT
 
-NAME = 'datagovsg'
-VERSION = '2.0.0' # Production
-VERSION = f'{VERSION}.{datetime.now().strftime("%Y%m%d%H%M")}' # Development
-AUTHOR = 'Yuhui'
-AUTHOR_EMAIL = 'yuhuibc@gmail.com'
+TRANSPORT_API_ENDPOINT = f'{BASE_V1_API_ENDPOINT}/transport'
+
+CARPARK_AVAILABILITY_API_ENDPOINT = \
+    f'{TRANSPORT_API_ENDPOINT}/carpark-availability'
 
 __all__ = [
-    'Ckan',
-    'Economy',
-    'Environment',
-    'Housing',
-    'Transport',
+    'CARPARK_AVAILABILITY_API_ENDPOINT',
 ]
-__version__ = VERSION
