+++ datagovsg/transport/constants.py
@@ -18,13 +18,10 @@ from ..constants import BASE_V1_API_ENDPOINT
 
 TRANSPORT_API_ENDPOINT = f'{BASE_V1_API_ENDPOINT}/transport'
 
-CARPARK_AVAILABILITY_API_ENDPOINT = \
-    f'{TRANSPORT_API_ENDPOINT}/carpark-availability'
 TAXI_AVAILABILITY_API_ENDPOINT = f'{TRANSPORT_API_ENDPOINT}/taxi-availability'
 TRAFFIC_IMAGES_API_ENDPOINT = f'{TRANSPORT_API_ENDPOINT}/traffic-images'
 
 __all__ = [
-    'CARPARK_AVAILABILITY_API_ENDPOINT',
     'TAXI_AVAILABILITY_API_ENDPOINT',
     'TRAFFIC_IMAGES_API_ENDPOINT',
 ]
