+++ datagovsg/transport/types.py
@@ -34,62 +34,6 @@ class _ApiInfoDict(TypedDict):
     - "healthy"
     """
 
-# Carpark Availability
-
-class _CarparkAvailabilityItemDataInfoDict(TypedDict):
-    """Type definition for _CarparkAvailabilityItemDataDict"""
-
-    total_lots: int
-    """Total number of carpark lots.
-
-    :example: 105
-    """
-    lot_type: str
-    """Type of carpark lot.
-
-    :example: "C"
-    """
-    lots_available: int
-    """Total number of available carpark lots.
-
-    :example: 86
-    """
-
-class _CarparkAvailabilityItemDataDict(TypedDict):
-    """Type definition for _CarparkAvailabilityItemDict"""
-
-    carpark_info: list[_CarparkAvailabilityItemDataInfoDict]
-    """Carpark information."""
-    carpark_number: str
-    """Carpark number.
-
-    :example: "HE12"
-    """
-    update_datetime: datetime
-    """Update date-time.
-
-    :example: datetime(2025, 1, 12, 15, 59, 0, \
-        tzinfo=zoneinfo.ZoneInfo(key='Asia/Singapore'))
-    """
-
-class _CarparkAvailabilityItemDict(TypedDict):
-    """Type definition for CarparkAvailabilityDict"""
-
-    timestamp: datetime
-    """Time of acquisition of data.
-
-    :example: datetime(2025, 1, 12, 15, 59, 0, \
-        tzinfo=zoneinfo.ZoneInfo(key='Asia/Singapore'))
-    """
-    carpark_data: list[_CarparkAvailabilityItemDataDict]
-    """Carpark availability information per carpark."""
-
-class CarparkAvailabilityDict(TypedDict):
-    """Type definition for carpark_availability()"""
-
-    items: list[_CarparkAvailabilityItemDict]
-    """Items."""
-
 # Taxi Availability
 
 class _TaxiAvailabilityCrsPropertiesDict(TypedDict):
